+++ Assignment_2/package.json
@@ -44,6 +44,7 @@
     "chart.js": "^2.6.0",
     "mobx": "^3.2.1",
     "mobx-react": "^4.2.2",
+    "mobx-react-devtools": "^4.2.15",
     "moment": "^2.18.1",
     "react": "^15.6.1",
     "react-chartjs-2": "^2.5.1",
