+++ Assignment_2/src/frontend/controller/WeatherPageContainer.tsx
@@ -1,12 +1,9 @@
 import * as React from 'react';
 import * as SocketIo from 'socket.io-client';
-
+import { observer } from 'mobx-react';
 import {
   FullLambdaServiceClient,
-  MonitorConnection,
-  OnLocationsRetrievedObserver,
-  OnMonitorAddedObserver,
-  OnWeatherLocationDataListRetrievedObserver,
+  MonitorConnection
 } from '../../lambda_client/FullLambdaServiceClient';
 
 import { AppState } from '../model/AppState';
@@ -20,8 +17,9 @@ import { WeatherLocationData } from '../../model/WeatherLocationData';
 import { WeatherPage } from '../view/WeatherPage';
 
 interface WeatherPageContainerProps {
-  readonly regularServiceUrl: string;
-  readonly timelapseServiceUrl: string;
+  readonly appState: AppState;
+  readonly regularClient: FullLambdaServiceClient;
+  readonly timelapseClient: FullLambdaServiceClient;
   readonly regularServicePrefix: string;
   readonly timelapseServicePrefix: string;
 }
@@ -33,107 +31,16 @@ interface WeatherPageContainerProps {
  * The class acts as a controller in that it coordinates the events of view objects and the interaction to the 
  * backend API.
  */
-class WeatherPageContainer extends React.Component<WeatherPageContainerProps, AppState> {
+@observer
+class WeatherPageContainer extends React.Component<WeatherPageContainerProps, {}> {
   private onLocationsListRainfallItemClicked: OnLocationItemClickedObserver;
   private onLocationsListTemperatureItemClicked: OnLocationItemClickedObserver;
   private onMonitoringListGraphItemClicked: OnMonitoringItemClickedObserver;
-  private readonly regularServiceClient: FullLambdaServiceClient;
-  private readonly timelapseServiceClient: FullLambdaServiceClient;
-
-  constructor(props: WeatherPageContainerProps) {
-    super(props);
-    // Start the state off with a bunch of empty lists.
-    this.state = new AppState([], new Map<string, MonitoredLocationInformation>(), false);
-    // Connects to the port that the backend is listening on.
-    // Triggers io.on('connection')'s callback
-    this.regularServiceClient = new FullLambdaServiceClient(SocketIo.connect(this.props.regularServiceUrl));
-    this.timelapseServiceClient = new FullLambdaServiceClient(SocketIo.connect(this.props.timelapseServiceUrl));
-  }
-  
-  // Create anon class to handle what happens when locations are retrieved.
-  private createOnLocationsRetrievedObserver(
-    servicePrefix: string, 
-    serviceTitle: string
-  ): OnLocationsRetrievedObserver {
-    // Set the locations and 
-    return (sortedLocations: string[]) => {
-        // Now that we have the locations, we need to initialize the MonitoredLocationInformation.
-        // You could lazy-initialize them but that would more complicated code with minimal benefits.
-        for (const location of sortedLocations) {
-          const prefixedLocation: string = LocationServicePrefixer.prefixLocation(servicePrefix, location);
-          this.state.weatherDataMap.set(
-            prefixedLocation, 
-            new MonitoredLocationInformation(
-              location,
-              serviceTitle, 
-              [], 
-              false, 
-              false, 
-              false
-            )
-          );
-          AppState.insertServiceLocation(this.state, servicePrefix, location);
-        }
-        console.log(this.state);
-    };
-  }
-
-  // Create anon class to handle retrieving a list of weather data.
-  private createWeatherDataListRetrievedObserver(servicePrefix: string): OnWeatherLocationDataListRetrievedObserver {
-    return (weatherLocationDataList: WeatherLocationData[]) => {
-        // We received some fresh weather data.
-        // Tell React that we may need to re-render
-        // Handle updates for cards and adding a new data point to graphs.
-        // Use to determine what cards are rendered and what information is in them (textual, graphical)
-        // and for rainfall and/or temperature.
-        const timeStamp: string = new Date().toString();
-        console.log('Received weather location data at time: ' + timeStamp);
-        console.log(weatherLocationDataList);
-          
-        const newWeatherDataMap: Map<string, MonitoredLocationInformation> = this.state.weatherDataMap;
-        console.log(this.state);
-        // Loop for each WeatherLocationData object sent by backend.
-        for (const weatherLocationData of weatherLocationDataList) {
-          const monitoredLocationInformation: MonitoredLocationInformation | undefined = newWeatherDataMap
-            .get(LocationServicePrefixer.prefixLocation(servicePrefix, weatherLocationData.location));
-          if (monitoredLocationInformation == null) {
-            throw new Error('No monitoring information was retrieved.');
-          }
-          // Add this weatherLocationData received to array of weatherLocationData.
-          monitoredLocationInformation.weatherDataList.push(weatherLocationData);
-        }
-    };
-  }
-
-  // Create anon class to handle adding a monitor response.
-  private createServiceMonitorAddedObserver(servicePrefix: string): OnMonitorAddedObserver {
-    return (addMonitorResponse: RequestResponse<WeatherLocationData>) => {
-        // First, make sure we didn't receive an error
-        if (addMonitorResponse.error == null) {
-          // Good, we didn't receive an error, add the new weather data into our state's weather hash map.
-          const newWeatherData: WeatherLocationData = addMonitorResponse.data;
-          const weatherDataMap: Map<string, MonitoredLocationInformation> = this.state.weatherDataMap;
-          const prefixedLocation: string = LocationServicePrefixer.prefixLocation(
-              servicePrefix, newWeatherData.location);
-          const monitoringData: MonitoredLocationInformation | undefined 
-            = weatherDataMap.get(prefixedLocation);
-          if (monitoringData != null) {
-            weatherDataMap.set(prefixedLocation, monitoringData);
-            monitoringData.weatherDataList.push(newWeatherData);
-          } else {
-            console.error('Could not find monitoring data');
-          }
-        } else {
-          console.error(addMonitorResponse.error);
-        }
-    };
-  }
-
   public componentDidMount(): void {
     // Create on click monitor listeners
     this.onMonitoringListGraphItemClicked = (locationKey: string) => {
         const monitoredLocationInformation: MonitoredLocationInformation | undefined = 
-          this.state.weatherDataMap.get(locationKey);
+          this.props.appState.weatherDataMap.get(locationKey);
         if (monitoredLocationInformation != null) { 
           const newMonitoredLocationInformation: MonitoredLocationInformation = new MonitoredLocationInformation(
             monitoredLocationInformation.location,
@@ -144,7 +51,7 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
             !monitoredLocationInformation.monitorGraph
           );
           // Update WeatherDataMap.
-          this.state.weatherDataMap.set(locationKey, newMonitoredLocationInformation);
+          this.props.appState.weatherDataMap.set(locationKey, newMonitoredLocationInformation);
         } else {
           console.error(`Error: monitoredLocationInformation could not be found for ${locationKey}`);
         }
@@ -156,7 +63,7 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
         // selected is the previous state, weather the button was previously selected or not.
         // If not selected before then selected will be false, we pass in !selected to make it true
         // so we render that component.
-        const originalData: MonitoredLocationInformation | undefined = this.state.weatherDataMap.get(prefixedLocation);
+        const originalData: MonitoredLocationInformation | undefined = this.props.appState.weatherDataMap.get(prefixedLocation);
         let newData: MonitoredLocationInformation;
         if (originalData == null) {
            throw new Error('There was no monitoring information.');
@@ -171,7 +78,7 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
         );
 
         // Add new data to the state in AppState weatherMap in memory.
-        this.state.weatherDataMap.set(prefixedLocation, newData);
+        this.props.appState.weatherDataMap.set(prefixedLocation, newData);
 
         this.onMonitorSelected(
           this.selectServiceClient(prefixedLocation).rainfallMonitorConnection, 
@@ -181,7 +88,7 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
     };
 
     this.onLocationsListTemperatureItemClicked = (prefixedLocation: string, selected: boolean) => {
-        const originalData: MonitoredLocationInformation | undefined = this.state.weatherDataMap.get(prefixedLocation);
+        const originalData: MonitoredLocationInformation | undefined = this.props.appState.weatherDataMap.get(prefixedLocation);
         let newData: MonitoredLocationInformation;
         if (originalData == null) {
           throw new Error('Could not find orginal monitoring information.');
@@ -194,18 +101,13 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
           !selected,
           originalData.monitorGraph
         );
-        this.state.weatherDataMap.set(prefixedLocation, newData);
+        this.props.appState.weatherDataMap.set(prefixedLocation, newData);
         this.onMonitorSelected(
           this.selectServiceClient(prefixedLocation).temperatureMonitorConnection,
           new MonitorMetadata(newData.location),
           selected
         );
     };
-
-    // Initialize the socket end points for the original FullLambdaService from stage 1.
-    this.initializeServiceClientObservers(this.regularServiceClient, this.props.regularServicePrefix, 'Original');
-    // Initialize the socket end points for the original FullLambdaService from stage 2.
-    this.initializeServiceClientObservers(this.timelapseServiceClient, this.props.timelapseServicePrefix, 'Timelapse');
   }
 
   /**
@@ -213,9 +115,9 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
    */
   private selectServiceClient(prefixedLocation: string): FullLambdaServiceClient {
     if (prefixedLocation.startsWith(this.props.regularServicePrefix)) {
-      return this.regularServiceClient;
+      return this.props.regularClient;
     } else if (prefixedLocation.startsWith(this.props.timelapseServicePrefix)) {
-      return this.timelapseServiceClient;
+      return this.props.timelapseClient;
     } else {
       throw new Error(`Could not select service client from prefixed location: ${prefixedLocation}`);
     }
@@ -238,40 +140,14 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
     }
   }
   
-  private initializeServiceClientObservers(
-    serviceClient: FullLambdaServiceClient,
-    servicePrefix: string,
-    serviceTitle: string
-  ): void {
-    // Create the weather data list observer
-    serviceClient.addOnWeatherLocationDataListRetrievedObserver(
-      this.createWeatherDataListRetrievedObserver(servicePrefix)
-    );
-    // Create the server setup observer.
-    serviceClient.addOnServerSetupSuccessRetrievedObserver(
-        (success: boolean) => {
-          this.state.connectedToServer = success;
-        }
-    );
-    serviceClient.addOnLocationsRetrievedObserver(this.createOnLocationsRetrievedObserver(servicePrefix, serviceTitle));
-    // Create observers specific to this service.
-    const onMonitorAddedObserver: OnMonitorAddedObserver = this.createServiceMonitorAddedObserver(servicePrefix);
-
-    // Rainfall monitors
-    serviceClient.rainfallMonitorConnection.addMonitorAddedObserver(onMonitorAddedObserver);
-
-    // Temperature monitors
-    serviceClient.temperatureMonitorConnection.addMonitorAddedObserver(onMonitorAddedObserver);
-    // TODO: We can reconfirm the monitorRainfall and monitorTemperature via removeMonitorEvent observers.
-  }
   
   public render(): JSX.Element {
     // console.log(this.state.weatherDataMap);
     return (
-      this.state.connectedToServer ?
+      this.props.appState.connectedToServer ?
       (
         <WeatherPage 
-          appCurrentState={this.state}
+          appCurrentState={this.props.appState}
           onLocationRainfallItemClickedObserver={this.onLocationsListRainfallItemClicked}
           onLocationTemperatureItemClickedObserver={this.onLocationsListTemperatureItemClicked}
           onMonitoringListGraphItemClicked={this.onMonitoringListGraphItemClicked}
@@ -289,6 +165,5 @@ class WeatherPageContainer extends React.Component<WeatherPageContainerProps, Ap
     );
   }
 }
-
 export {WeatherPageContainer};
 export default WeatherPageContainer;
