+++ Assignment_2/tsconfig.frontend.json
@@ -17,7 +17,7 @@
     "noUnusedLocals": true,
     "experimentalDecorators": true,
     "emitDecoratorMetadata": true,
-    "noEmitHelpers": true 
+    "noEmitHelpers": false 
   },
   "files": [],
   "exclude": [
